feat(standards): add get-elements-by-content-type and implicit-html-roles (#2375)
* feat(standards): add get-elements-by-content-type and implicit-html-roles

* cleanup

* lookuptable to use implicit roles from commons/standards
