+++ lib/commons/aria/implicit-role.js
@@ -1,4 +1,4 @@
-import lookupTable from './lookup-table';
+import implicitHtmlRoles from '../standards/implicit-html-roles';
 import { getNodeFromTree } from '../../core/utils';
 import AbstractVirtuaNode from '../../core/base/virtual-node/abstract-virtual-node';
 
@@ -32,7 +32,7 @@ function implicitRole(node) {
 	}
 
 	const nodeName = vNode.props.nodeName;
-	const role = lookupTable.implicitHtmlRole[nodeName];
+	const role = implicitHtmlRoles[nodeName];
 
 	if (!role) {
 		return null;
