+++ lib/commons/standards/index.js
@@ -7,3 +7,5 @@ export { default as getAriaRolesByType } from './get-aria-roles-by-type';
 export { default as getAriaRolesSupportingNameFromContent } from './get-aria-roles-supporting-name-from-content';
 export { default as getGlobalAriaAttrs } from './get-global-aria-attrs';
 export { default as getElementSpec } from './get-element-spec';
+export { default as getElementsByContentType } from './get-elements-by-content-type';
+export { default as implicitHtmlRoles } from './implicit-html-roles';
